experiments/cnn_sliding_window/requirements_cnn.txt
# CNN Sliding Window Experiments - Full Requirements
# (Archived - for reference only)

# Core dependencies
opencv-python>=4.8.0
jupyter>=1.0.0
matplotlib>=3.7.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
tensorflow>=2.15.0

# ML pipeline and experiment tracking
mlflow>=2.9.0
tensorboard>=2.15.0

# Hyperparameter tuning
keras-tuner>=1.4.6

# Configuration management
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality (optional but recommended)
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
